pythonProject/requirements.txt
mysql-connector-python
wxPython